src-django/requirements.txt
Django==1.9.3
Fabric==1.10.2
bcrypt==2.0.0
celery==3.1.20
django-appconf==1.0.1
django-compressor==2.0
django-cors-headers==1.1.0
django-extensions==1.6.1
django-filter==0.12.0
django-nose==1.4.3
django-postgres-copy==0.0.7
django-redis-cache==1.6.5
djangorestframework==3.3.2
factory-boy==2.6.1
flake8==2.5.4
freezegun==0.3.6
gunicorn==19.4.5
jinja2==2.8
mock==1.3.0
nose==1.3.7
psycopg2==2.6.1
redis==2.10.5
pyfcm==1.4.4
